pyproject.toml
[project]
name = "ascii-3d-surface-renderer"
version = "0.1.0"
description = "Renders 3D ASCII Surfaces to terminal."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "sympy>=1.14.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "single"
